sync github issues
