+++ .github/workflows/api-tracker.yaml
@@ -1,4 +1,4 @@
-name: Prefect API tracker
+name: Track API Parity and create issues
 
 'on':
   # Run weekly on Monday at 9:00 AM UTC
@@ -23,6 +23,7 @@ jobs:
       contents: read
       # required for the paths-filter GHA, if workflow triggered by PR
       pull-requests: read
+      issues: write
     runs-on: ubuntu-latest
     steps:
       - name: Checkout repository
@@ -30,46 +31,12 @@ jobs:
         with:
           fetch-depth: 0  # We need history to compare with previous schemas
 
-      - name: Get previous API schema
-        uses: actions/download-artifact@v4
+      - name: Install uv
+        uses: astral-sh/setup-uv@v5
         with:
-          name: openapi.json
+          version: "0.6.13"
 
-      - name: Compare with latest schema
-        run: |
-          /bin/bash ./scripts/compare.sh
-
-      - name: Check if diff found
-        id: diff
-        uses: andstor/file-existence-action@v3
-        with:
-          files: 'diff.md'
-
-      - name: Create GitHub issue for changes
-        if: steps.diff.outputs.files_exists == 'true'
-        run: |
-          gh issue create \
-            --title='Prefect API change detected' \
-            --body-file='diff.md' \
-            --label='maintenance'
-
-      - name: Prepare current schema for upload
-        run: |
-          # Remove the previous schema file
-          rm openapi.json
-          # Rename the current schema file
-          mv openapi-current.json openapi.json
-
-      - name: Upload current schema for future comparison
-        if: steps.diff.outputs.files_exists == 'true'
-        uses: actions/upload-artifact@v4
-        with:
-          name: openapi.json
-          path: openapi.json
-          overwrite: true
-          # The flow will run every 7 days, so no need to
-          # hold on to the file for much longer than that.
-          retention-days: 10
-          # We want to make sure the current schema is uploaded
-          # at the end to generate proper incremental diffs.
-          if-no-files-found: error
+      - name: Run compare.py
+        run: uv run scripts/compare-and-sync-issues.py
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
